# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Gestión de Biblioteca",
    "description": "API para registrar, listar, editar y eliminar.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://192.168.1.8:5000"
    }
  ],
...
  "paths": {
    "/usuarios": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Registrar un nuevo usuario",
        "description": "Este endpoint permite registrar un nuevo usuario en el sistema. Requiere enviar un objeto JSON con los datos del usuario como nombres,apellidos, correo,telefono,direccion y contraseña.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "primer_nombre",
                  "primer_apellido",
                  "correo",
                  "telefono",
                  "direccion",
                  "contraseña"
                ],
                "properties": {
                  "primer_nombre": {
                    "type": "string"
                  },
                  "primer_apellido": {
                    "type": "string"
                  },
                  "correo": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "contrasenia": {
                    "type": "string"
                  }
                },
                "example": {
                  "primer_nombre": "diana",
                  "primer_apellido": "ortiz",
                  "correo": "danaortiz0711@gmail.com",
                  "telefono": "3043306181",
                  "direccion": "soledad",
                  "contrasenia": "Diana2006$"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje de confirmación indicando que el usuario fue registrado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Usuario Registrado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "faltan parametro en la peticion..."
                  }
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Correo Invalido"
                  }
                }
              }
            }
          },
          "406": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Contraseña Invalida"
                  }
                }
              }
            }
          }
        }
      }
      ,
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Listar todos los Usuarios",
        "description": "Este endpoint permite obtener un listado completo de los Usuarios registrados en el sistema. La respuesta incluye información básica como id,nombres,apellidos,estado, correo,telefono,direccion y contraseña.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta exitosa que contiene la lista de aprendices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "primer_nombre",
                      "primer_apellido",
                      "estado",
                      "correo",
                      "telefono",
                      "direccion",
                      "contraseña"
                    ],
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "primer_nombre": {
                        "type": "string"
                      },
                      "primer_apellido": {
                        "type": "string"
                      },
                      "estado": {
                        "type": "integer"
                      },
                      "correo": {
                        "type": "string"
                      },
                      "telefono": {
                        "type": "string"
                      },
                      "direccion": {
                        "type": "string"
                      },
                      "contrasenia": {
                        "type": "string"
                      }
                    },
                    "example": {
                      "id": "6",
                      "primer_nombre": "diana",
                      "primer_apellido": "ortiz",
                      "correo": "danaortiz0711@gmail.com",
                      "estado": "1",
                      "telefono": "3043306181",
                      "direccion": "soledad",
                      "contrasenia": "Diana2006$"
                    }
                  }
                }
              }
            }
          }
        }
      }

    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Iniciar sesion",
        "description": "Permite iniciar sesion al usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "correo",
                  "contraseña"
                ],
                "properties": {
                  "correo": {
                    "type": "string"
                  },
                  "contrasenia": {
                    "type": "string"
                  }
                },
                "example": {
                  "correo": "danaortiz0711@gmail.com",
                  "contrasenia": "Diana2006$"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje de confirmación indicando que el usuario fue logueado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Iniciando sesion"
                  }
                }
              }
            }
          },
          "401": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "El correo no esta registrado"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{uid}": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Editar usuario",
        "parameters": [
          {
            "name": "uid",
            "in": "put",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Este endpoint permite editar un  usuario en el sistema. Requiere enviar un objeto JSON con los datos del usuario como nombres,apellidos, correo,telefono,direccion.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "primer_nombre",
                  "primer_apellido",
                  "correo",
                  "telefono",
                  "direccion"
                ],
                "properties": {
                  "primer_nombre": {
                    "type": "string"
                  },
                  "primer_apellido": {
                    "type": "string"
                  },
                  "correo": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  }
                },
                "example": {
                  "primer_nombre": "diana",
                  "primer_apellido": "ortiz",
                  "correo": "danaortiz0711@gmail.com",
                  "telefono": "3043306181",
                  "direccion": "soledad"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Devuelve un mensaje de confirmación indicando que el usuario fue registrado correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Usuario Actualizado"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "faltan parametro en la peticion..."
                  }
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Usuario no encontrado"
                  }
                }
              }
            }
          },
          "422": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Correo Invalido"
                  }
                }
              }
            }
          },
          "409": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "El correo electrónico ya está registrado"
                  }
                }
              }
            }
          }
        }
      }
      ,
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Eliminar un Usuario por ID",
        "description": "Este endpoint permite eliminar un Usuario del sistema utilizando su identificador único (ID). Si el ID proporcionado es válido y corresponde a un Usuario existente, se eliminará de forma permanente.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve un mensaje de confirmación indicando que el Usuario fue eliminado correctamente del sistema.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "Usuario Eliminado"
                  }
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "mensaje": "No se encontro ningun usuario"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}